+++ cmd/enduro/main.go
@@ -208,13 +208,13 @@ func main() {
 		)
 	}
 
-	// Set up upload bucket.
-	internalBucket, err := cfg.InternalStorage.OpenBucket(ctx)
+	// Set up internal storage bucket.
+	internalStorage, err := cfg.InternalStorage.OpenBucket(ctx)
 	if err != nil {
-		logger.Error(err, "Error setting up internal bucket.")
+		logger.Error(err, "Error setting up internal storage.")
 		os.Exit(1)
 	}
-	defer internalBucket.Close()
+	defer internalStorage.Close()
 
 	// Set up the ingest service.
 	var ingestsvc ingest.Service
@@ -228,7 +228,7 @@ func main() {
 			tokenVerifier,
 			ticketProvider,
 			cfg.Temporal.TaskQueue,
-			internalBucket,
+			internalStorage,
 			cfg.Upload.MaxSize,
 			rand.Reader,
 		)
@@ -319,7 +319,7 @@ func main() {
 			&auth.NoopTokenVerifier{},
 			ticketProvider,
 			cfg.Temporal.TaskQueue,
-			internalBucket,
+			internalStorage,
 			cfg.Upload.MaxSize,
 			rand.Reader,
 		)
