+++ enduro.toml
@@ -210,10 +210,12 @@ passphrase = "" # Secret: set (if required) with env var ENDURO_AM_SFTP_PRIVATEK
 # Default: 4294967296 (4 GiB).
 maxSize = 4294967296
 
-# internalBucket section configures a bucket where Enduro will place uploaded
+# internalStorage section configures a bucket where Enduro will place uploaded
 # SIPs. Make sure it doesn't match any of the watched buckets.
 [internalStorage.bucket]
-# URL = "azblob://sips" # Example for an Azure URL connection where the bucket/container is called sips
+# Example for an Azure URL connection where the bucket/container is called "sips".
+# See [internalStorage.azure] below to set credentials.
+# URL = "azblob://sips"
 endpoint = "http://minio.enduro-sdps:9000"
 pathStyle = true
 accessKey = "minio"
